Refactor to add interactive test environment
Create run_fakeserver.py intended to be run with an interactive python
shell for development purposes. This entailed splitting our overridden
implementation of MCRcon along with related classes used in the test
framework so they could be shared.

This commit modifies run.py to launch this interactive fake_server
in a tmux session.