Catch an occasional error on fake_server teardown
There was still an odd occasional exception being thrown for tests
using fake_server where the server was not available for an RCON
connection even though the object returned is_serving was true. I
think there are just some odd timing issues associated with the
asyncio-driven fake_app and as the fake_app is not my primary goal
I don't want to dedicate the time to deep dive in hopes of fixing
things in there.